Renamed properties to model
