+++ src/main/java/model/RoomDetails.java
@@ -1,4 +1,4 @@
-package Properties;
+package model;
 
 public class RoomDetails {
     String roomType;
